requirements.txt
google-generativeai
numpy
openpyxl
pandas
Pillow
python-dotenv
streamlit
